rename PageLayout to SettingLayout
