+++ src/renderer/components/+add-cluster/add-cluster.tsx
@@ -38,8 +38,8 @@ import { navigate } from "../../navigation";
 import { iter } from "../../utils";
 import { AceEditor } from "../ace-editor";
 import { Button } from "../button";
-import { PageLayout } from "../layout/page-layout";
 import { Notifications } from "../notifications";
+import { SettingLayout } from "../layout/setting-layout";
 
 interface Option {
   config: KubeConfig;
@@ -108,7 +108,7 @@ export class AddCluster extends React.Component {
 
   render() {
     return (
-      <PageLayout className="AddClusters" showOnTop={true}>
+      <SettingLayout className="AddClusters">
         <h2>Add Clusters from Kubeconfig</h2>
         <p>
           Clusters added here are <b>not</b> merged into the <code>~/.kube/config</code> file.
@@ -144,7 +144,7 @@ export class AddCluster extends React.Component {
             tooltipOverrideDisabled
           />
         </div>
-      </PageLayout>
+      </SettingLayout>
     );
   }
 }
