+++ src/renderer/components/+extensions/extensions.tsx
@@ -39,12 +39,12 @@ import logger from "../../../main/logger";
 import { Button } from "../button";
 import { ConfirmDialog } from "../confirm-dialog";
 import { DropFileInput, InputValidators } from "../input";
-import { PageLayout } from "../layout/page-layout";
 import { Notifications } from "../notifications";
 import { ExtensionInstallationState, ExtensionInstallationStateStore } from "./extension-install.store";
 import { Install } from "./install";
 import { InstalledExtensions } from "./installed-extensions";
 import { Notice } from "./notice";
+import { SettingLayout } from "../layout/setting-layout";
 
 function getMessageFromError(error: any): string {
   if (!error || typeof error !== "object") {
@@ -317,7 +317,7 @@ export async function attemptInstallByInfo({ name, version, requireConfirmation 
         version = json["dist-tags"][version];
       } else {
         Notifications.error(<p>The <em>{name}</em> extension does not have a version or tag <code>{version}</code>.</p>);
-  
+
         return disposer();
       }
     }
@@ -510,7 +510,7 @@ export class Extensions extends React.Component<Props> {
 
     return (
       <DropFileInput onDropFiles={installOnDrop}>
-        <PageLayout showOnTop className="Extensions" contentGaps={false}>
+        <SettingLayout className="Extensions" contentGaps={false}>
           <section>
             <h1>Extensions</h1>
 
@@ -533,7 +533,7 @@ export class Extensions extends React.Component<Props> {
               uninstall={confirmUninstallExtension}
             />
           </section>
-        </PageLayout>
+        </SettingLayout>
       </DropFileInput>
     );
   }
