+++ src/renderer/components/+preferences/preferences.tsx
@@ -31,7 +31,6 @@ import { AppPreferenceRegistry, RegisteredAppPreference } from "../../../extensi
 import { UserStore } from "../../../common/user-store";
 import { ThemeStore } from "../../theme.store";
 import { Input } from "../input";
-import { PageLayout } from "../layout/page-layout";
 import { SubTitle } from "../layout/sub-title";
 import { Select, SelectOption } from "../select";
 import { HelmCharts } from "./helm-charts";
@@ -40,6 +39,7 @@ import { navigation } from "../../navigation";
 import { Tab, Tabs } from "../tabs";
 import { FormSwitch, Switcher } from "../switch";
 import { KubeconfigSyncs } from "./kubeconfig-syncs";
+import { SettingLayout } from "../layout/setting-layout";
 
 enum Pages {
   Application = "application",
@@ -136,8 +136,7 @@ export class Preferences extends React.Component {
       );
 
     return (
-      <PageLayout
-        showOnTop
+      <SettingLayout
         navigation={this.renderNavigation()}
         className="Preferences"
         contentGaps={false}
@@ -266,7 +265,7 @@ export class Preferences extends React.Component {
             {extensions.filter(e => !e.showInPreferencesTab).map(this.renderExtension)}
           </section>
         )}
-      </PageLayout>
+      </SettingLayout>
     );
   }
 }
