+++ src/renderer/components/layout/setting-layout.scss
@@ -19,12 +19,19 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-.PageLayout {
+.SettingLayout {
   --width: 75%;
   --nav-width: 180px;
   --nav-column-width: 30vw;
 
-  position: relative;
+  position: fixed !important; // allow to cover Hotbar
+  z-index: 13;
+  left: 0;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  height: unset;
+  background-color: var(--settingsBackground);
   width: 100%;
   height: 100%;
   display: grid !important;
@@ -40,18 +47,6 @@
     > .contentRegion {
       justify-content: flex-start;
     }
-  }
-
-  // covers whole app view area
-  &.showOnTop {
-    position: fixed !important; // allow to cover ClustersMenu
-    z-index: 13;
-    left: 0;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    height: unset;
-    background-color: var(--settingsBackground);
   }
 
   > .sidebarRegion {
